Fix JDBC-39 by forcing Boolean column extraction when metadata indicates BOOLEAN type.
Since (.getBoolean) seems able to sometimes return a Boolean false that isn't identical to CLojure's false, the result of (.getBoolean) is forced to nil, true or false as appropriate.
Additional tests have been added for MySQL only. I have not been able to reproduce the not-a-boolean situation however, to verify that calling the (pure-boolean) function is appropriate.

This forms the basis for extracting other columns based on their correct, known SQL type, regardless of other field flags.