db-do-prepared* now take [sql param param] or multi? JDBC-122
`db-do-prepared` and `db-do-prepared-return-keys` now accept both the
legacy style of arguments (db sql param-group param-group ...) and the
new style (db [sql param param] {} or db [sql param-group] {:multi?
true}) so they are compatible in style with the rest of the library.

The 0.5.0 tests all still pass (and produce lots of DEPRECATED
warnings. The 0.5.8 tests all pass (and produce NO deprecated warnings).

I'm committing this as a functional milestone while I refactor and clean
up the implementation of a few functions.