Fix JDBC-131 by ignoring clean up exceptions
If the clean up fails after a transaction is complete (regardless of
whether the transaction succeeded or failed), ignore those failures and
continue with remaining cleanup.

The downside of this is that a connection may silently get into an
unexpected state if we can't clean it up (but the most likely cause is
that the connection itself became corrupt or was closed so it is almost
certainly unusable anyway and the next operation will probably fail
too!).