Tweaks to readme to highlight docs
Even knowing where things are I always have trouble finding the links to the API reference and docs. My aim with these changes was to make it easier to find, to streamline some of the info, and hopefully get more people to go to the docs at clojure-doc.org.
These changes make sense to me, but that doesn't mean they will make sense to other people or be desired - I'm happy to make changes to this PR or an outright rejection if my changes aren't desirable.